requirements.txt
aiogram
environs
tortoise-orm
asyncpg
python-telegram-bot-pagination